requirements.txt
confluent-kafka==2.6.1
py4j==0.10.9.7
pyspark==3.5.3
python-dotenv==1.0.1
